WebCL-conformance/conformance/extension/WEBCL_html_image/cl_commandQueue_enqueueWriteBuffer_with_HTMLImageElement.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWriteBuffer with HTMLImageElement conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    var image = new Image();
    image.src = "../../../resources/pebbleHalf.jpg";
    image.onload = execute;
} catch(e) {
    testFailed("Unable to load HTML image element : " + e.name);
}
var webCLContext, webCLCommandQueue, webCLBufferRead, webCLBufferWrite, webCLBufferReadWrite, bufferOffset = 0, bufferOffsetInvalid, webCLBuffer;
var bufferSize, releasedBuffer, bufferSizeSmall, webCLBufferSmall, webCLEvent;
var webCLPlatform, devices, webCLDevice1, webCLDevice2, webCLContext1, webCLContext2, webCLCommandQueue1, webCLCommandQueue2, webCLBuffer1, webCLBuffer2;
var releasedEvent, webCLUserEvent, webCLUserEvent1, webCLUserEvent2, webCLEvent1, webCLEvent2, invalid_image;
function execute() {
    try {
        debug("WebCLCommandQueue :: enqueueWriteBuffer() with HTMLImageElement test case.");
        debug("(WebCLBuffer buffer, CLboolean blockingWrite, CLuint bufferOffset, HTMLImageElement srcImage, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

        var wtu = WebCLTestUtils;
        webCLContext = wtu.createContext();
        webCLCommandQueue = wtu.createCommandQueue(webCLContext);
        bufferSize = image.width * image.height * 4;

        webCLBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBuffer, true, bufferOffset, image)", "WEBCL_EXTENSION_NOT_ENABLED");
        if (!wtu.enableExtension(webcl, "WEBCL_html_image")) {
            throw { description : "Failed to enable extension : WEBCL_html_image. Exitting..." };
        }

        wtu.releaseAll(webCLContext);
        webCLContext = wtu.createContext();
        webCLCommandQueue = wtu.createCommandQueue(webCLContext);

        // MEM_READ_ONLY Buffers.
        webCLBufferRead = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferRead, true, bufferOffset, image)");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferRead, false, bufferOffset, image)");

        // MEM_WRITE_ONLY Buffers.
        webCLBufferWrite = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, bufferSize);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferWrite, true, bufferOffset, image)");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferWrite, false, bufferOffset, image)");

        // MEM_READ_WRITE Buffers.
        webCLBufferReadWrite = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSize);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image)");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, false, bufferOffset, image)");

        // Testing for nullable parameters.
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, [])");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, [], null)");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, null)");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, null, null)");
        webCLEvent = wtu.createEvent();
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, null, webCLEvent)");
        webCLEvent = wtu.createEvent();
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, [], webCLEvent)");

        bufferOffsetNonZero = 10;
        bufferSizeBig = (image.width * image.height * 4) + bufferOffsetNonZero;
        webCLBufferBig = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSizeBig);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferBig, true, bufferOffsetNonZero, image)");

        // Testing for exceptions.
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffsetNonZero, image)", "INVALID_VALUE");

        releasedBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSize);
        wtu.release(releasedBuffer);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(releasedBuffer, true, bufferOffset, image)", "INVALID_MEM_OBJECT");
        invalid_image = new Image();
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, invalid_image)", "INVALID_HOST_PTR");

        bufferOffsetInvalid = bufferSize + 1;
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffsetInvalid, image)", "INVALID_VALUE");

        bufferSizeSmall = (image.width * image.height * 4) - 1;
        webCLBufferSmall = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSizeSmall);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferSmall, true, bufferOffset, image)", "INVALID_VALUE");

        // Testing for exception "INVALID_CONTEXT"
        webCLPlatform = wtu.getPlatform();
        devices = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ALL);
        if (devices.length > 1) {
            webCLDevice1 = devices[0];
            webCLContext1 = wtu.createContext(webCLDevice1);
            webCLCommandQueue1 = wtu.createCommandQueue(webCLContext1, webCLDevice1);
            webCLBuffer1 = wtu.createBuffer(webCLContext1, webcl.MEM_READ_WRITE, bufferSize);

            webCLDevice2 = devices[1];
            webCLContext2 = wtu.createContext(webCLDevice2);
            webCLCommandQueue2 = wtu.createCommandQueue(webCLContext2, webCLDevice2);
            webCLBuffer2 = wtu.createBuffer(webCLContext2, webcl.MEM_READ_WRITE, bufferSize);

            shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, image)", "INVALID_CONTEXT");
            shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, image)", "INVALID_CONTEXT");

            webCLUserEvent1 = wtu.createUserEvent(webCLContext1);
            webCLUserEvent2 = wtu.createUserEvent(webCLContext2);
            shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, false, bufferOffset, image, [webCLUserEvent2])", "INVALID_CONTEXT");
            shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, false, bufferOffset, image, [webCLUserEvent1])", "INVALID_CONTEXT");
            wtu.setStatus(webCLUserEvent1, webcl.COMPLETE);
            wtu.setStatus(webCLUserEvent2, webcl.COMPLETE);

            webCLEvent1 = wtu.createEvent();
            webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, image, [], webCLEvent1);
            shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, image, [webCLEvent1])", "INVALID_CONTEXT");
            webCLEvent2 = wtu.createEvent();
            webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, image, [], webCLEvent2);
            shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, image, [webCLEvent2])", "INVALID_CONTEXT");
        }

        releasedEvent = wtu.createUserEvent(webCLContext);
        wtu.release(releasedEvent);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, [releasedEvent])", "INVALID_EVENT_WAIT_LIST");

        webCLUserEvent = wtu.createUserEvent(webCLContext);
        wtu.setStatus(webCLUserEvent, -1);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, image, [webCLUserEvent])", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

    } catch (e) {
        testFailed(e.description);
    }

    wtu.appendPostJSToHTML(document);
}
successfullyParsed = true;
</script>
</body>
</html>
